requirements.txt
Cheetah==2.4.4
Django==1.4.8
Fabric==1.4.0
Markdown==2.1.1
PIL==1.1.7
Paste==1.7.5.1
PasteDeploy==1.5.0
PasteScript==1.7.5
South==0.7.6
Werkzeug==0.8.3
boto==2.8.0
dj-database-url==0.2.1
django-appconf==0.6
django-compressor==1.3
django-debug-toolbar==0.9.4
django-extensions==0.9
django-filter==0.7
django-guardian==1.1.0.beta
django-heroku-memcacheify==0.4
django-project-templates==0.11
django-pylibmc-sasl==0.2.4
django-registration==0.8
-e git://github.com/jamstooks/django-s3-folder-storage.git@153cf2a462dad27703b09ca477f3e014a5c03e47#egg=django_s3_folder_storage-0.1-py2.7-dev
django-storages==1.1.6
django-taggit==0.9.3
-e hg+https://bitbucket.org/fabian/django-taggit-autosuggest@ce0382212840c05fd98a12782ea636051fdd04e6#egg=django_taggit_autosuggest-tip
django-tastypie==0.9.12
geopy==0.94.2
greenlet==0.3.4
gunicorn==0.13.4
html2text==3.200.3
html5lib==0.99
mimeparse==0.1.3
newrelic==1.10.2.38
psycopg2==2.5
pyPdf==1.13
pycrypto==2.5
python-dateutil==1.5
python-documentcloud==0.154
pytidylib==0.2.1
reportlab==2.6
requests==1.1.0
simplejson==3.1.0
six==1.4.1
ssh==1.7.13
wsgiref==0.1.2
pytz==2013.7
xhtml2pdf==0.0.4
xlrd==0.9.0
pylibmc==1.2.3
MySQL-python==1.2.3
django-sslify==0.2
workdays==1.2
django-bleach==0.1.5
